Conflicts:
	libstd/bld.sh

+++ libstd/bld.sh
@@ -55,19 +55,22 @@ MYR="types.myr \
 
 OBJ="$(echo $ASM | sed 's/\.s/.o /g') $(echo $MYR | sed 's/\.myr/.o /g')"
 USE="$(echo $MYR | sed 's/\.myr/.use /g' | sed "s/-$SYS//g")"
-rm -f $OBJ test libstd.a
-assem $ASM
-use $MYR
-build $MYR
+if [ "$1" = "clean" ]; then
+    echo rm -f $OBJ test libstd.a
+    rm -f $OBJ test libstd.a
+else
+    assem $ASM
+    use $MYR
+    build $MYR
 
-echo $MU -mo std $USE
-$MU -mo std $USE
-echo ar -rcs libstd.a $OBJ
-ar -rcs libstd.a $OBJ
+    echo $MU -mo std $USE
+    $MU -mo std $USE
+    echo ar -rcs libstd.a $OBJ
+    ar -rcs libstd.a $OBJ
 
-# build test program
-build test.myr
-COMP="$LD -o test test.o -L. -lstd"
-echo $COMP
-$COMP
-
+    # build test program
+    build test.myr
+    COMP="$LD -o test test.o -L. -lstd"
+    echo $COMP
+    $COMP
+fi
